requirements.txt
amqp==1.4.6
anyjson==0.3.3
billiard==3.3.0.20
celery==3.1.18
Django==1.8.2
djangorestframework==3.4.7
kombu==3.0.30
pkg-resources==0.0.0
pytz==2015.4
redis==2.10.3
requests==2.7.0
